move binding helpers to its main folder
